refactor: Featured Event 컴포넌트 분리
